+++ src/lib/components/CameraControls.svelte
@@ -3,7 +3,7 @@
 	import { CameraControls, type CameraControlsRef, Gizmo, Portal } from '@threlte/extras'
 	import { useCameraControls, useTransformControls } from '$lib/hooks/useControls.svelte'
 	import KeyboardControls from './KeyboardControls.svelte'
-	import Button from './dashboard/Button.svelte'
+	import Button from '$lib/components/overlay/dashboard/Button.svelte'
 	import { useSettings } from '$lib/hooks/useSettings.svelte'
 
 	const cameraControls = useCameraControls()