+++ src/lib/entries.ts
@@ -1,7 +1,0 @@
-export const typeSafeObjectFromEntries = <
-	const T extends ReadonlyArray<readonly [PropertyKey, unknown]>,
->(
-	entries: T
-): { [K in T[number] as K[0]]: K[1] } => {
-	return Object.fromEntries(entries) as { [K in T[number] as K[0]]: K[1] }
-}