+++ src/lib/frame.ts
@@ -14,7 +14,6 @@ type FrameGeometryMap = {
 }
 
 export type FrameGeometry = keyof FrameGeometryMap
-export type FrameGeometries = ValueOf<FrameGeometryMap>
 
 type FrameOrientationMap = {
 	quaternion: { type: 'quaternion'; value: { x: number; y: number; z: number; w: number } }
@@ -24,7 +23,6 @@ type FrameOrientationMap = {
 }
 
 export type FrameOrientation = keyof FrameOrientationMap
-export type FrameOrientations = ValueOf<FrameOrientationMap>
 
 export interface Frame<
 	T extends FrameGeometry = FrameGeometry,