+++ src/lib/test.ts
@@ -1,1 +1,0 @@
-export { createRandomPcdBinary } from './test/createRandomPcdBinary'